refactor(switchboard,reactor-api): single OTel+Sentry bootstrap
Three separate bootstrap sites were racing for the same OTel globals on
switchboard startup:

- packages/reactor-api/src/tracing.ts started a NodeSDK at module-load
  (side-effect of importing reactor-api/server.ts)
- apps/switchboard/src/metrics.ts built a standalone MeterProvider passed
  in via a startServer option that then called setGlobalMeterProvider
- apps/switchboard/src/server.mts called Sentry.init, whose @sentry/node
  default registers its own OTel TracerProvider

Whichever ran last clobbered setGlobalTracerProvider/setGlobalMeterProvider.

Move all three into apps/switchboard/src/observability.mts, imported as
the very first module in index.mts so OpenTelemetry require-time hooks
register before http/express/pg/graphql load. reactor-api becomes a pure
library: tracing.ts deleted, the trace() pass-through helper inlined at
its four call sites, the meterProvider plumbing removed.

createMeterProviderFromEnv stays in apps/switchboard/src/metrics.ts so
its test suite (12 cases against undocumented MeterProvider internals)
keeps passing; observability.mts composes it.

OTel deps moved from reactor-api/package.json to apps/switchboard/package.json
(instrumentation-{express,graphql,http,pg}, sdk-node, sdk-trace-base,
exporter-trace-otlp-http, semantic-conventions). @opentelemetry/exporter-prometheus
dropped — the PROMETHEUS_METRICS_PORT pull path in the old tracing.ts had
no in-tree consumer.

Verified with scripts/profiling/reactor-direct.ts --otel: same 10 metric
families emitted, total time 8.01s → 8.01s, p99 6ms → 6ms.
