+++ apps/switchboard/src/index.mts
@@ -1,8 +1,12 @@
 #!/usr/bin/env node
+// Observability MUST load before any module that imports http/express/pg/graphql
+// so OpenTelemetry's require-time hooks can patch them. It also owns Sentry
+// init and the SIGINT/SIGTERM flush.
+import "./observability.mjs";
+
 import * as Sentry from "@sentry/node";
 import { childLogger } from "document-model";
 import { config } from "./config.js";
-import { createMeterProviderFromEnv } from "./metrics.js";
 import { initProfilerFromEnv } from "./profiler.js";
 import { startSwitchboard } from "./server.mjs";
 
@@ -28,27 +32,6 @@ ensureNodeVersion("24");
 // scales with dynamically-loaded document models beyond the default cap of 10.
 process.setMaxListeners(0);
 
-const meterProvider = createMeterProviderFromEnv({
-  OTEL_EXPORTER_OTLP_ENDPOINT: process.env.OTEL_EXPORTER_OTLP_ENDPOINT,
-  OTEL_METRIC_EXPORT_INTERVAL: process.env.OTEL_METRIC_EXPORT_INTERVAL,
-  OTEL_SERVICE_NAME: process.env.OTEL_SERVICE_NAME,
-});
-
-async function shutdown() {
-  console.log("\nShutting down...");
-  // Flush final metrics before exit. Races against a 5s deadline so an
-  // unresponsive OTLP endpoint cannot exhaust terminationGracePeriodSeconds.
-  await Promise.race([
-    meterProvider?.shutdown().catch(() => undefined),
-    new Promise<void>((resolve) => setTimeout(resolve, 5_000)),
-  ]);
-  process.exit(0);
-}
-
-// SIGINT: Ctrl-C in development; SIGTERM: graceful shutdown in Docker/Kubernetes
-process.on("SIGINT", shutdown);
-process.on("SIGTERM", shutdown);
-
 if (process.env.PYROSCOPE_SERVER_ADDRESS) {
   try {
     await initProfilerFromEnv(process.env);
@@ -64,5 +47,4 @@ startSwitchboard({
   ...config,
   migratePglite: cliMigratePglite || config.migratePglite,
   forcePgVersion: config.forcePgVersion ?? undefined,
-  meterProvider,
 }).catch(console.error);
