+++ packages/reactor-api/index.mts
@@ -8,6 +8,5 @@ export * from "./src/services/auth.service.js";
 export * from "./src/services/document-permission.service.js";
 export * from "./src/services/package-management.service.js";
 export * from "./src/services/package-storage.js";
-export * from "./src/tracing.js";
 export * from "./src/types.js";
 export * from "./src/utils/index.js";
