+++ packages/reactor-api/package.json
@@ -51,17 +51,6 @@
     "@apollo/subgraph": "^2.13.2",
     "@as-integrations/express4": "^1.1.2",
     "@electric-sql/pglite": "catalog:",
-    "@opentelemetry/instrumentation-express": "^0.48.0",
-    "@opentelemetry/instrumentation-graphql": "^0.48.0",
-    "@opentelemetry/instrumentation-http": "^0.200.0",
-    "@opentelemetry/instrumentation-pg": "^0.52.0",
-    "@opentelemetry/exporter-metrics-otlp-http": "^0.57.2",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.57.2",
-    "@opentelemetry/resources": "^1.29.0",
-    "@opentelemetry/sdk-metrics": "^1.29.0",
-    "@opentelemetry/sdk-node": "^0.57.2",
-    "@opentelemetry/sdk-trace-base": "^1.29.0",
-    "@opentelemetry/semantic-conventions": "^1.29.0",
     "@originjs/vite-plugin-commonjs": "^1.0.3",
     "@powerhousedao/analytics-engine-core": "workspace:*",
     "@powerhousedao/analytics-engine-graphql": "workspace:*",
@@ -100,7 +89,6 @@
     "read-pkg": "catalog:",
     "ws": "^8.18.3",
     "zod": "catalog:",
-    "@opentelemetry/exporter-prometheus": "^0.57.2",
     "graphql-tag": "catalog:"
   },
   "devDependencies": {
